♻️ refactor: Update go modules and fix gohack undo in pre-commit
This commit updates the go modules to the latest versions, including:
- Upgrading `github.com/shirou/gopsutil/v4` from `v4.25.4` to `v4.25.5`
- Updating several indirect dependencies

Additionally, it fixes an issue in the pre-commit configuration where the `gohack undo` command was not adding the `go.mod` file after undoing the hacks. This commit ensures that the `go.mod` file is added after running `gohack undo`.